1.Entity层和Mapper层完善：增加了User类，Teacher类等，完善了实体类设计。并且完善了相应mapper类的设计

+++ backend/src/main/java/com/example/aicourse/entity/Course.java
@@ -3,6 +3,7 @@ package com.example.aicourse.entity;
 import com.baomidou.mybatisplus.annotation.*;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data @TableName("t_course")
@@ -12,7 +13,7 @@ public class Course{
  private Long id;
 
  private String courseCode,courseName;
- private Double credit;
+ private BigDecimal credit;
  private Integer hours;
  private Long teacherId;
